seem to have been fixed

+++ src/chem/y/chem/equilibrium.cpp
@@ -19,7 +19,7 @@ namespace upsylon
 
         }
 
-        int Equilibrium::Component::Compare(const Component *lhs, const Component *rhs,void*)
+        ptrdiff_t Equilibrium::Component::Compare(const Component *lhs, const Component *rhs,void*)
         {
             return comparison::increasing(lhs->sp->indx,rhs->sp->indx);
         }
