+++ src/graphic/y/graphic/ops/edges/build.cpp
@@ -8,7 +8,7 @@ namespace upsylon {
 
 
         static inline
-        int compareEdges( const Edge *lhs, const Edge *rhs, void * ) throw()
+        ptrdiff_t compareEdges( const Edge *lhs, const Edge *rhs, void * ) throw()
         {
             return comparison::decreasing(lhs->size,rhs->size);
         }
